refactor(LmsNavigationField): create
